fix(ci): full-coverage maintainability — split oversized provider_readiness part_002 under 1500-line limit
Move test_cli_auth_probe_reports_success_and_nonzero_failure and
test_redaction_parts_cover_urls_empty_secrets_existing_redactions_and_merges
from part_002 (1507 lines) into the existing part_003, bringing part_002
to 1456 lines so test_first_party_code_files_stay_under_line_limit passes.
