+++ tests/unit/service/test_provider_readiness_parts/test_provider_readiness_part_002.py
@@ -1454,54 +1454,3 @@ def test_agent_runtime_cli_probe_reports_structured_failures(
     assert result["status"] == "fail"
     assert result["reason_code"] == reason_code
     assert "sk-runtime-secret" not in json.dumps(result, sort_keys=True)
-
-
-@pytest.mark.unit
-def test_cli_auth_probe_reports_success_and_nonzero_failure() -> None:
-    ok = provider_readiness._probe_cli_auth_status(
-        provider_label="Codex",
-        args=["codex", "auth", "status"],
-        failure_reason="CODEX_AUTH_FAILED",
-        timeout_reason="CODEX_AUTH_TIMEOUT",
-        missing_reason="CODEX_AUTH_CLI_MISSING",
-        error_reason="CODEX_AUTH_ERROR",
-        environ={},
-        run_subprocess=lambda _args, **_kwargs: _completed(stdout="ok\n"),
-        secrets=frozenset(),
-    )
-    failed = provider_readiness._probe_cli_auth_status(
-        provider_label="Codex",
-        args=["codex", "auth", "status"],
-        failure_reason="CODEX_AUTH_FAILED",
-        timeout_reason="CODEX_AUTH_TIMEOUT",
-        missing_reason="CODEX_AUTH_CLI_MISSING",
-        error_reason="CODEX_AUTH_ERROR",
-        environ={},
-        run_subprocess=lambda _args, **_kwargs: _completed(returncode=1, stderr="bad auth"),
-        secrets=frozenset(),
-    )
-
-    assert ok == {"status": "ok", "reason_code": "CODEX_AUTH_OK"}
-    assert failed["status"] == "fail"
-    assert failed["reason_code"] == "CODEX_AUTH_FAILED"
-    assert failed["detail"] == "bad auth"
-
-
-@pytest.mark.unit
-def test_redaction_parts_cover_urls_empty_secrets_existing_redactions_and_merges() -> None:
-    redacted, parts = provider_readiness._redact_with_redaction_parts(
-        "connect http://user:pass@example.test with sk-providersecret123456",
-        frozenset({""}),
-    )
-
-    assert "user:pass" not in redacted
-    assert "sk-providersecret123456" not in redacted
-    assert parts is not None
-    assert provider_readiness_helpers._slice_redaction_segments(
-        [("literal", "before"), ("redaction", ""), ("literal", "after")],
-        6,
-        16,
-    ) == [("redaction", "")]
-    assert provider_readiness_helpers._merge_literal_redaction_segments(
-        [("literal", ""), ("literal", "a"), ("literal", "b")]
-    ) == [("literal", "ab")]
